fix non-compliant build integration tests
